fix(platform): P2-2 统一 timeout 类型为 Int64 ms
- 所有 ATimeoutMs 参数统一为 Int64 类型
  - console.pas: platform_console_wait_readable
  - io.pas: platform_poller_wait
  - process.pas: platform_process_wait
  - watch.pas: platform_watch_poll
  - socket.pas: platform_socket_set_timeout (AMs: UInt32 → ATimeoutMs: Int64)
- sync.pas: ATimeoutNs: Int64 保持不变（纳秒精度是同步原语的刚需）
- thread.pas: ATimeoutMs: Int64 已经是正确类型
- 更新 CONTRACT.md 和 USABILITY-ASSESSMENT.md
- 所有 33 个测试套件通过，0 unfreed

+++ core/src/nextpas.core.platform.console.pas
@@ -57,7 +57,7 @@ function platform_console_write(AFd: Int32; ABuf: Pointer; ACount: Int32): Int32
  * @return cwReadable / cwTimeout / cwInterrupted（被信号打断）/ cwError
  * @note 不在内部重试 EINTR——调用方据此消费信号 pending 状态后自行决定重试。
  *}
-function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int32): TPlatformConsoleWait;
+function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int64): TPlatformConsoleWait;
 
 implementation
 
@@ -176,7 +176,7 @@ begin
   Result := LSent;
 end;
 
-function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int32): TPlatformConsoleWait;
+function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int64): TPlatformConsoleWait;
 var
   LPfd: pollfd;
   LRc: cint;
@@ -206,7 +206,7 @@ function platform_console_read(AFd: Int32; ABuf: Pointer; ACount: Int32): Int32;
 begin Result := PLATFORM_ERR_UNSUPPORTED; end;
 function platform_console_write(AFd: Int32; ABuf: Pointer; ACount: Int32): Int32;
 begin Result := PLATFORM_ERR_UNSUPPORTED; end;
-function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int32): TPlatformConsoleWait;
+function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int64): TPlatformConsoleWait;
 begin Result := cwError; end;
 {$ENDIF}
 {$ENDIF}
@@ -369,7 +369,7 @@ begin
   Result := LSent;
 end;
 
-function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int32): TPlatformConsoleWait;
+function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int64): TPlatformConsoleWait;
 var
   LHandle: HANDLE;
   LTimeout: DWORD;
@@ -409,7 +409,7 @@ function platform_console_read(AFd: Int32; ABuf: Pointer; ACount: Int32): Int32;
 begin Result := PLATFORM_ERR_UNSUPPORTED; end;
 function platform_console_write(AFd: Int32; ABuf: Pointer; ACount: Int32): Int32;
 begin Result := PLATFORM_ERR_UNSUPPORTED; end;
-function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int32): TPlatformConsoleWait;
+function platform_console_wait_readable(AFd: Int32; ATimeoutMs: Int64): TPlatformConsoleWait;
 begin Result := cwError; end;
 {$ENDIF}
